chore

+++ src/main/java/com/example/crawler/meal/repository/MenuRepository.java
@@ -1,10 +1,7 @@
 package com.example.crawler.meal.repository;
 
 import com.example.crawler.meal.entity.Menu;
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
-public interface MenuRepository extends JpaRepository<Menu, Long> {
-
+public interface MenuRepository extends JpaRepository<Menu, Integer> {
 }
